add zarr router
- Move all the zarr logic from `rest.py` to `routers/zarr.py`.

- Wrap the zarr layer in a new (internal) Dataset accessor for easy
  reuse in path operation functions.
