+++ xpublish/routers/__init__.py
@@ -1,2 +1,3 @@
 from .base import base_router, get_dataset
 from .common import common_router
+from .zarr import zarr_router
